requirements.txt
feedparser
newspaper3k
youtube-transcript-api
pyyaml
requests
beautifulsoup4
markdownify
python-dotenv
lxml_html_clean
langchain
langchain-community
ollama
pydantic
requests